Move census code to production_statistics.rs
